+++ src/program/MovieFile.h
@@ -104,7 +104,7 @@ public:
     uint64_t savestateFramecount() const;
 
     /* Get the movie length from metadata */
-    void length(uint64_t* sec, uint64_t* nsec) const;
+    void length(int64_t* sec, int64_t* nsec) const;
 
     /* Set inputs for a certain frame, and truncate if keep_inputs is false */
     int setInputs(const AllInputs& inputs, uint64_t pos, bool keep_inputs);